#197 Fix collisions for inner classes with same name
