+++ reflective-fluent-builders-generator/src/main/java/io/github/tobi/laa/reflective/fluent/builders/generator/impl/BuilderClassNameGeneratorImpl.java
@@ -6,6 +6,8 @@ import io.github.tobi.laa.reflective.fluent.builders.model.BuilderMetadata;
 
 import javax.inject.Named;
 import javax.inject.Singleton;
+import java.util.ArrayDeque;
+import java.util.Deque;
 import java.util.Objects;
 
 /**
@@ -20,6 +22,24 @@ class BuilderClassNameGeneratorImpl implements BuilderClassNameGenerator {
     @Override
     public ClassName generateClassName(final BuilderMetadata builderMetadata) {
         Objects.requireNonNull(builderMetadata);
-        return ClassName.get(builderMetadata.getPackageName(), builderMetadata.getName());
+        final var builderHierarchy = enclosingBuilderHierarchy(builderMetadata);
+        var currentBuilder = builderHierarchy.pop();
+        var className = ClassName.get(currentBuilder.getPackageName(), currentBuilder.getName());
+        while (!builderHierarchy.isEmpty()) {
+            currentBuilder = builderHierarchy.pop();
+            className = className.nestedClass(currentBuilder.getName());
+        }
+        return className;
+    }
+
+    private Deque<BuilderMetadata> enclosingBuilderHierarchy(final BuilderMetadata builderMetadata) {
+        final Deque<BuilderMetadata> builderHierarchy = new ArrayDeque<>();
+        builderHierarchy.push(builderMetadata);
+        var currentEnclosingBuilder = builderMetadata.getEnclosingBuilder();
+        while (currentEnclosingBuilder.isPresent()) {
+            builderHierarchy.push(currentEnclosingBuilder.get());
+            currentEnclosingBuilder = currentEnclosingBuilder.get().getEnclosingBuilder();
+        }
+        return builderHierarchy;
     }
 }
