+++ reflective-fluent-builders-maven-plugin/src/it/resources/expected-builders/default-config/io/github/tobi/laa/reflective/fluent/builders/test/models/nested/TopLevelClassBuilder.java
@@ -96,4 +96,277 @@ public class TopLevelClassBuilder {
 
     TopLevelClass.NestedPublicLevelOne nestedPublic;
   }
+
+  @Generated(
+      value = "io.github.tobi.laa.reflective.fluent.builders.generator.api.JavaFileGenerator",
+      date = "3333-03-13T00:00Z[UTC]"
+  )
+  public static class NestedPackagePrivateLevelOneBuilder {
+    /**
+     * This field is solely used to be able to detect generated builders via reflection at a later stage.
+     */
+    @SuppressWarnings("unused")
+    private boolean ______generatedByReflectiveFluentBuildersGenerator;
+
+    private final Supplier<TopLevelClass.NestedPackagePrivateLevelOne> objectSupplier;
+
+    private final CallSetterFor callSetterFor = new CallSetterFor();
+
+    private final FieldValue fieldValue = new FieldValue();
+
+    protected NestedPackagePrivateLevelOneBuilder(
+        final Supplier<TopLevelClass.NestedPackagePrivateLevelOne> objectSupplier) {
+      this.objectSupplier = Objects.requireNonNull(objectSupplier);
+    }
+
+    public static NestedPackagePrivateLevelOneBuilder newInstance() {
+      return new NestedPackagePrivateLevelOneBuilder(TopLevelClass.NestedPackagePrivateLevelOne::new);
+    }
+
+    public static NestedPackagePrivateLevelOneBuilder withSupplier(
+        final Supplier<TopLevelClass.NestedPackagePrivateLevelOne> supplier) {
+      return new NestedPackagePrivateLevelOneBuilder(supplier);
+    }
+
+    public NestedPackagePrivateLevelOneBuilder field(final int field) {
+      this.fieldValue.field = field;
+      this.callSetterFor.field = true;
+      return this;
+    }
+
+    public TopLevelClass.NestedPackagePrivateLevelOne build() {
+      final TopLevelClass.NestedPackagePrivateLevelOne objectToBuild = this.objectSupplier.get();
+      if (this.callSetterFor.field) {
+        objectToBuild.setField(this.fieldValue.field);
+      }
+      return objectToBuild;
+    }
+
+    private class CallSetterFor {
+      boolean field;
+    }
+
+    private class FieldValue {
+      int field;
+    }
+  }
+
+  @Generated(
+      value = "io.github.tobi.laa.reflective.fluent.builders.generator.api.JavaFileGenerator",
+      date = "3333-03-13T00:00Z[UTC]"
+  )
+  public static class NestedProtectedLevelOneBuilder {
+    /**
+     * This field is solely used to be able to detect generated builders via reflection at a later stage.
+     */
+    @SuppressWarnings("unused")
+    private boolean ______generatedByReflectiveFluentBuildersGenerator;
+
+    private final Supplier<TopLevelClass.NestedProtectedLevelOne> objectSupplier;
+
+    private final CallSetterFor callSetterFor = new CallSetterFor();
+
+    private final FieldValue fieldValue = new FieldValue();
+
+    protected NestedProtectedLevelOneBuilder(
+        final Supplier<TopLevelClass.NestedProtectedLevelOne> objectSupplier) {
+      this.objectSupplier = Objects.requireNonNull(objectSupplier);
+    }
+
+    public static NestedProtectedLevelOneBuilder newInstance() {
+      return new NestedProtectedLevelOneBuilder(TopLevelClass.NestedProtectedLevelOne::new);
+    }
+
+    public static NestedProtectedLevelOneBuilder withSupplier(
+        final Supplier<TopLevelClass.NestedProtectedLevelOne> supplier) {
+      return new NestedProtectedLevelOneBuilder(supplier);
+    }
+
+    public NestedProtectedLevelOneBuilder field(final int field) {
+      this.fieldValue.field = field;
+      this.callSetterFor.field = true;
+      return this;
+    }
+
+    public TopLevelClass.NestedProtectedLevelOne build() {
+      final TopLevelClass.NestedProtectedLevelOne objectToBuild = this.objectSupplier.get();
+      if (this.callSetterFor.field) {
+        objectToBuild.setField(this.fieldValue.field);
+      }
+      return objectToBuild;
+    }
+
+    private class CallSetterFor {
+      boolean field;
+    }
+
+    private class FieldValue {
+      int field;
+    }
+  }
+
+  @Generated(
+      value = "io.github.tobi.laa.reflective.fluent.builders.generator.api.JavaFileGenerator",
+      date = "3333-03-13T00:00Z[UTC]"
+  )
+  public static class NestedPublicLevelOneBuilder {
+    /**
+     * This field is solely used to be able to detect generated builders via reflection at a later stage.
+     */
+    @SuppressWarnings("unused")
+    private boolean ______generatedByReflectiveFluentBuildersGenerator;
+
+    private final Supplier<TopLevelClass.NestedPublicLevelOne> objectSupplier;
+
+    private final CallSetterFor callSetterFor = new CallSetterFor();
+
+    private final FieldValue fieldValue = new FieldValue();
+
+    protected NestedPublicLevelOneBuilder(
+        final Supplier<TopLevelClass.NestedPublicLevelOne> objectSupplier) {
+      this.objectSupplier = Objects.requireNonNull(objectSupplier);
+    }
+
+    public static NestedPublicLevelOneBuilder newInstance() {
+      return new NestedPublicLevelOneBuilder(TopLevelClass.NestedPublicLevelOne::new);
+    }
+
+    public static NestedPublicLevelOneBuilder withSupplier(
+        final Supplier<TopLevelClass.NestedPublicLevelOne> supplier) {
+      return new NestedPublicLevelOneBuilder(supplier);
+    }
+
+    public NestedPublicLevelOneBuilder nested(
+        final TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo nested) {
+      this.fieldValue.nested = nested;
+      this.callSetterFor.nested = true;
+      return this;
+    }
+
+    public TopLevelClass.NestedPublicLevelOne build() {
+      final TopLevelClass.NestedPublicLevelOne objectToBuild = this.objectSupplier.get();
+      if (this.callSetterFor.nested) {
+        objectToBuild.setNested(this.fieldValue.nested);
+      }
+      return objectToBuild;
+    }
+
+    private class CallSetterFor {
+      boolean nested;
+    }
+
+    private class FieldValue {
+      TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo nested;
+    }
+
+    @Generated(
+        value = "io.github.tobi.laa.reflective.fluent.builders.generator.api.JavaFileGenerator",
+        date = "3333-03-13T00:00Z[UTC]"
+    )
+    public static class NestedPublicLevelTwoBuilder {
+      /**
+       * This field is solely used to be able to detect generated builders via reflection at a later stage.
+       */
+      @SuppressWarnings("unused")
+      private boolean ______generatedByReflectiveFluentBuildersGenerator;
+
+      private final Supplier<TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo> objectSupplier;
+
+      private final CallSetterFor callSetterFor = new CallSetterFor();
+
+      private final FieldValue fieldValue = new FieldValue();
+
+      protected NestedPublicLevelTwoBuilder(
+          final Supplier<TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo> objectSupplier) {
+        this.objectSupplier = Objects.requireNonNull(objectSupplier);
+      }
+
+      public static NestedPublicLevelTwoBuilder newInstance() {
+        return new NestedPublicLevelTwoBuilder(TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo::new);
+      }
+
+      public static NestedPublicLevelTwoBuilder withSupplier(
+          final Supplier<TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo> supplier) {
+        return new NestedPublicLevelTwoBuilder(supplier);
+      }
+
+      public NestedPublicLevelTwoBuilder nested(
+          final TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree nested) {
+        this.fieldValue.nested = nested;
+        this.callSetterFor.nested = true;
+        return this;
+      }
+
+      public TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo build() {
+        final TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo objectToBuild = this.objectSupplier.get();
+        if (this.callSetterFor.nested) {
+          objectToBuild.setNested(this.fieldValue.nested);
+        }
+        return objectToBuild;
+      }
+
+      private class CallSetterFor {
+        boolean nested;
+      }
+
+      private class FieldValue {
+        TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree nested;
+      }
+
+      @Generated(
+          value = "io.github.tobi.laa.reflective.fluent.builders.generator.api.JavaFileGenerator",
+          date = "3333-03-13T00:00Z[UTC]"
+      )
+      public static class NestedPublicLevelThreeBuilder {
+        /**
+         * This field is solely used to be able to detect generated builders via reflection at a later stage.
+         */
+        @SuppressWarnings("unused")
+        private boolean ______generatedByReflectiveFluentBuildersGenerator;
+
+        private final Supplier<TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree> objectSupplier;
+
+        private final CallSetterFor callSetterFor = new CallSetterFor();
+
+        private final FieldValue fieldValue = new FieldValue();
+
+        protected NestedPublicLevelThreeBuilder(
+            final Supplier<TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree> objectSupplier) {
+          this.objectSupplier = Objects.requireNonNull(objectSupplier);
+        }
+
+        public static NestedPublicLevelThreeBuilder newInstance() {
+          return new NestedPublicLevelThreeBuilder(TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree::new);
+        }
+
+        public static NestedPublicLevelThreeBuilder withSupplier(
+            final Supplier<TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree> supplier) {
+          return new NestedPublicLevelThreeBuilder(supplier);
+        }
+
+        public NestedPublicLevelThreeBuilder field(final int field) {
+          this.fieldValue.field = field;
+          this.callSetterFor.field = true;
+          return this;
+        }
+
+        public TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree build(
+            ) {
+          final TopLevelClass.NestedPublicLevelOne.NestedPublicLevelTwo.NestedPublicLevelThree objectToBuild = this.objectSupplier.get();
+          if (this.callSetterFor.field) {
+            objectToBuild.setField(this.fieldValue.field);
+          }
+          return objectToBuild;
+        }
+
+        private class CallSetterFor {
+          boolean field;
+        }
+
+        private class FieldValue {
+          int field;
+        }
+      }
+    }
+  }
 }
